add touchhandler, move scroll methods to scroll handler
